work ok subgradient

+++ src/gap.h
@@ -25,7 +25,7 @@ int gap_branch_and_bound ();
 int gap_lower_bound ();
 
 //calculates the value of the lagrangian function
-int gap_calcuate_lagrangian_function (Problem * problem);
+double gap_calcuate_lagrangian_function (Problem * problem);
 
 //calculates initial solution for first relaxation
 void gap_calculate_initial_lagrangian (Problem * problem);
